requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
fastapi==0.93.0
filelock==3.9.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
pipenv==2023.2.18
platformdirs==3.1.0
pydantic==1.10.5
pymongo==4.3.3
python-dotenv==0.21.1
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
zipp==3.15.0
